clean up datatree branch
