+++ podaac/subsetter/tree_time_converting.py
@@ -1,5 +1,20 @@
+"""
+Converts the time variable to datetime if xarray doesn't decode times
+
+Parameters
+----------
+dataset : xr.Dataset
+time_vars : list
+instrument_type: string
+
+Returns
+-------
+xr.Dataset
+datetime.datetime
+"""
+
 import datetime
-from typing import Tuple, Union, Dict
+from typing import Tuple, Union
 import xarray as xr
 import numpy as np
 
@@ -32,14 +47,14 @@ def get_start_date(instrument_type: str) -> Union[datetime.datetime, None]:
 def get_group_by_path(data_tree: xr.Dataset, path: str) -> xr.Dataset:
     """
     Navigate through the data tree to get the group containing the variable
-    
+
     Parameters
     ----------
     data_tree : xr.Dataset
         The root dataset containing nested groups
     path : str
         Path to the variable in format '/group1/group2/var'
-        
+
     Returns
     -------
     xr.Dataset
@@ -49,11 +64,11 @@ def get_group_by_path(data_tree: xr.Dataset, path: str) -> xr.Dataset:
     """
     parts = path.strip('/').split('/')
     current_group = data_tree
-    
+
     # Navigate through all parts except the last (which is the variable name)
     for group_name in parts[:-1]:
         current_group = current_group[group_name]
-    
+
     return current_group, parts[-1]
 
 
@@ -82,7 +97,7 @@ def convert_to_datetime(data_tree: xr.Dataset, time_vars: list, instrument_type:
 
     for var_path in time_vars:
         group, var_name = get_group_by_path(data_tree, var_path)
-        
+
         if np.issubdtype(group[var_name].dtype, np.dtype(float)) or np.issubdtype(group[var_name].dtype, np.float32):
             # adjust the time values from the start date
             if start_date:
@@ -91,13 +106,13 @@ def convert_to_datetime(data_tree: xr.Dataset, time_vars: list, instrument_type:
                 # add seconds since the start time to the start time to get the time at the data point
                 group[var_name].values = date_time_array.astype("datetime64[ns]") + group[var_name].astype('timedelta64[s]').values
                 continue
-            
+
             # if there isn't a start_date, get it from the UTC variable
             utc_var_name = compute_utc_name(group)
             if utc_var_name:
                 start_seconds = group[var_name].values[0]
-                group[var_name].values = [datetime.datetime(i[0], i[1], i[2], hour=i[3], minute=i[4], second=i[5]) 
-                                        for i in group[utc_var_name].values]
+                group[var_name].values = [datetime.datetime(i[0], i[1], i[2], hour=i[3], minute=i[4], second=i[5])
+                                          for i in group[utc_var_name].values]
                 start_date = group[var_name].values[0] - np.timedelta64(int(start_seconds), 's')
                 return data_tree, start_date
         else:
